+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.15] - 2021-04-09
+### Added [#16](https://github.com/rafamizes/flutter_brand_palettes/issues/16)
+- Material Design's Pink color palette.
+- Material Design's Pink Accent color palette.
+
 ### Fixed
 - Fix typos in README file.